+++ scripts/deploy_dag.py
@@ -17,6 +17,7 @@ import argparse
 import pathlib
 import subprocess
 import typing
+import warnings
 
 CURRENT_PATH = pathlib.Path(__file__).resolve().parent
 PROJECT_ROOT = CURRENT_PATH.parent
@@ -27,6 +28,7 @@ def main(
     local: bool,
     env_path: pathlib.Path,
     dataset_id: str,
+    pipeline: str = None,
     airflow_home: pathlib.Path = None,
     composer_env: str = None,
     composer_bucket: str = None,
@@ -41,7 +43,12 @@ def main(
     )
 
     print("========== AIRFLOW DAGS ==========")
-    for pipeline_path in list_subdirs(env_path / "datasets" / dataset_id):
+    if pipeline:
+        pipelines = [env_path / "datasets" / pipeline]
+    else:
+        pipelines = list_subdirs(env_path / "datasets" / dataset_id)
+
+    for pipeline_path in pipelines:
         copy_custom_callables_to_airflow_dags_folder(
             local,
             env_path,
@@ -61,6 +68,10 @@ def main(
         )
 
 
+def run_gsutil_cmd(args: typing.List[str], cwd: pathlib.Path):
+    subprocess.check_call(["gsutil"] + args, cwd=cwd)
+
+
 def copy_variables_to_airflow_data_folder(
     local: bool,
     env_path: pathlib.Path,
@@ -74,7 +85,9 @@ def copy_variables_to_airflow_data_folder(
     cwd = env_path / "datasets" / dataset_id
     filename = f"{dataset_id}_variables.json"
 
-    check_existence_of_variables_file(cwd / filename)
+    if not (cwd / filename).exists():
+        warnings.warn(f"Airflow variables file {filename} does not exist.")
+        return
 
     if local:
         """
@@ -100,7 +113,31 @@ def copy_variables_to_airflow_data_folder(
             f"  Source:\n  {cwd / filename}\n\n"
             f"  Destination:\n  {gcs_uri}\n"
         )
-        subprocess.check_call(["gsutil", "cp", filename, gcs_uri], cwd=cwd)
+        run_gsutil_cmd(["cp", filename, gcs_uri], cwd=cwd)
+
+
+def run_cloud_composer_vars_import(
+    composer_env: str,
+    composer_region: str,
+    airflow_path: pathlib.Path,
+    cwd: pathlib.Path,
+):
+    subprocess.check_call(
+        [
+            "gcloud",
+            "composer",
+            "environments",
+            "run",
+            str(composer_env),
+            "--location",
+            str(composer_region),
+            "variables",
+            "--",
+            "--import",
+            str(airflow_path),
+        ],
+        cwd=cwd,
+    )
 
 
 def import_variables_to_airflow_env(
@@ -130,21 +167,8 @@ def import_variables_to_airflow_env(
         gcs_uri = f"gs://{composer_bucket}/data/variables/{filename}"
         airflow_path = f"/home/airflow/gcs/data/variables/{filename}"
         print(f"\nImporting Airflow variables from {gcs_uri} ({airflow_path})...\n")
-        subprocess.check_call(
-            [
-                "gcloud",
-                "composer",
-                "environments",
-                "run",
-                str(composer_env),
-                "--location",
-                str(composer_region),
-                "variables",
-                "--",
-                "--import",
-                str(airflow_path),
-            ],
-            cwd=cwd,
+        run_cloud_composer_vars_import(
+            composer_env, composer_region, airflow_path, cwd=cwd
         )
 
 
@@ -183,7 +207,7 @@ def copy_generated_dag_to_airflow_dags_folder(
             f"  Source:\n  {cwd / filename}\n\n"
             f"  Destination:\n  {target}\n"
         )
-        subprocess.check_call(["gsutil", "cp", filename, target], cwd=cwd)
+        run_gsutil_cmd(["cp", filename, target], cwd=cwd)
 
 
 def copy_custom_callables_to_airflow_dags_folder(
@@ -225,7 +249,7 @@ def copy_custom_callables_to_airflow_dags_folder(
             f"  Source:\n  {cwd / 'custom'}\n\n"
             f"  Destination:\n  {target}\n"
         )
-        subprocess.check_call(["gsutil", "cp", "-r", "custom", target], cwd=cwd)
+        run_gsutil_cmd(["cp", "-r", "custom", target], cwd=cwd)
 
 
 def check_existence_of_variables_file(file_path: pathlib.Path):
@@ -326,6 +350,7 @@ if __name__ == "__main__":
         local=args.local,
         env_path=PROJECT_ROOT / f".{args.env}",
         dataset_id=args.dataset,
+        pipeline=args.pipeline,
         airflow_home=airflow_path,
         composer_env=args.composer_env,
         composer_bucket=args.composer_bucket,