+++ scripts/generate_terraform.py
@@ -14,7 +14,6 @@
 
 
 import argparse
-import copy
 import pathlib
 import subprocess
 import typing
@@ -120,7 +119,7 @@ def generate_dataset_tf(dataset_id: str, project_id: str, config: dict, env: str
 
     contents = ""
     for resource in config["resources"]:
-        contents += tf_resource_contents(resource, copy.deepcopy(subs))
+        contents += tf_resource_contents(resource, {**resource, **subs})
 
     create_file_in_dir_tree(
         dataset_id, contents, f"{dataset_id}_dataset.tf", PROJECT_ROOT / f".{env}"
@@ -204,6 +203,14 @@ def customize_template_subs(resource: dict, subs: dict) -> dict:
         subs["uniform_bucket_level_access"] = resource.get(
             "uniform_bucket_level_access"
         )
+    elif resource["type"] == "bigquery_table":
+        # Terraform resource names cannot start with digits, but BigQuery allows
+        # table names that start with digits. We prepend `bqt_` to table names
+        # that doesn't comply with Terraform's naming rule.
+        if resource["table_id"][0].isdigit():
+            subs["tf_resource_name"] = "bqt_" + resource["table_id"]
+        else:
+            subs["tf_resource_name"] = resource["table_id"]
     return subs